Implement polyfills; fix enter & exit linking.
Polyfills are now implemented for mouseenter events, mouseleave events,
element.matches and CustomEvent. The CustomEvent polyfill is implemented lazily
on a per-window basis, whereas the others are only implemented once for the
global document (and thus are typically only available in browsers, and not
within the Node environment).

In addition, this commit changes how the update selection is linked to the enter
and update selections. Rather than linking each group, only the top-level update
selection object now has a reference to the enter selection and exit selection.
(Likewise, the enter selection now has a reference back to the update selection,
rather than linking from each enter group.)