Use a parallel array to store parent nodes.
Instead of decorating each group with a _parent, keep a parallel array of
parents. This is typically more efficient, and has the nice property that a
parents array can be shared when creating a subselection, since the parents are
immutable. (It’s also easier to test and inspect.)

Note that this means you can no longer access nodes._parent in a selection
function, but this was a private API.