Change enter.append(…, null) to append after the last child.
Previously it behaved the same as enter.append(…), such that entering nodes were
inserted before their following node in the update selection. This made it
impossible to use enter.append to insert entering nodes as the last child, as
selection.append(…) does! So now if a before node is explicitly specified, we
observe it even if it is null.