Tests for selection.merge.
Also relax the constraint that the two parents arrays must be the same instance.
Although this should be the common case when merging enter and update, it
doesn’t seem desirable to enforce this strictly; instead, we now define the
behavior that the returned selection always has the same structure (and parents)
as this selection, and that any null nodes in this selection may be populated
with the corresponding node from the other selection.