Don’t ignore duplicate keys.
Instead of ignoring duplicate keys, we now define symmetric behavior: if there
are multiple elements with the same key, the duplicates are added to the exit
selection; if there are multiple data with the same key, the duplicates are
added to the enter selection.

Fixes d3/d3-axis#10. Related d3/d3#997 d3/d3#999.