Use inline style value, if present.
Rather than always returning the computed style value, selection.style now
favors the inline value, if present, and only falls back to the computed value
if an inline value is not present.

By not computing the style value in the common case where an inline style is
present, this is faster, but more importantly it is more predictable, since a
computed value can be surprisingly different from an inline value, such as when
the inline value is specified as a percentage.

This change is especially useful for transition.style, d3/d3-transition#47, so
that you can control the starting style value of a transition.